+++ CHANGELOG.md
@@ -5,7 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+ 
+## [0.2.1] - 2018-10-25
+### Added
+- Change account key use case
 
+### Changed
+- Add `version` field for connection transactions
+
+### Fixed
+- Display connection transactions on a message view
+ 
 ## [0.2.0] - 2018-09-28
 ### Added
 - Import data starting from the last block
@@ -51,7 +61,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Contributing
 
 
-[Unreleased]: https://github.com/adshares/ads-operator/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/adshares/ads-operator/compare/v0.2.1...HEAD
 
+[0.2.1]: https://github.com/adshares/ads-operator/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/adshares/ads-operator/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/adshares/ads-operator/releases/tag/v0.1.0